Extract Renderer interface
