+++ Sources/XiEditor/XiTextPlane/GlRenderer/solid.f.glsl
