+++ Sources/XiEditor/XiTextPlane/GlRenderer/solid.v.glsl
