+++ Sources/XiEditor/XiTextPlane/GlRenderer/text.f.glsl
