+++ Sources/XiEditor/XiTextPlane/GlRenderer/text.v.glsl
